organized files properly as per features based folder structure
